Editor code cleanup & refactor

+++ app/src/main/java/hr/jkacan/setmaker/models/editor/EditorModels.kt
@@ -3,10 +3,10 @@ package hr.jkacan.setmaker.models.editor
 import hr.jkacan.setmaker.models.song.Song
 
 data class UiNode(
-    val id: Long,
+    val id: Int,
     val col: Int,   // lane (x)
     val row: Int,    // layer (y)
     val song: Song
 )
 
-data class UiEdge(val fromId: Long, val toId: Long)
+data class UiEdge(val fromId: Int, val toId: Int)
